dom/base/test/test_html_colors_quirks.html

<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=121738
-->
<head>
  <meta charset="UTF-8">
  <title>Test for Bug 121738</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=121738">Mozilla Bug 121738</a>
<table id="table0"></table>
<table id="table1"></table>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 121738 **/

String.prototype.toAsciiLowerCase = function () {
  var output = "";
  for (var i = 0, len = this.length; i < len; ++i) {
    if (this.charCodeAt(i) >= 0x41 && this.charCodeAt(i) <= 0x5A) {
      output += String.fromCharCode(this.charCodeAt(i) + 0x20)
    } else {
      output += this.charAt(i);
    }
  }
  return output;
}

var tests = [
"#135",
" #135",
"#135 ",
" #135 ",
"# 135",
"#   135",
"#123456",
" #123456",
"#123456 ",
" #123456 ",
"# 123456",
"#   123456",
"aliceblue",
"ALICEBLUE",
"alıceblue",
"alİceblue",
" aliceblue",
"aliceblue ",
"  aliceblue   ",
"antiquewhite",
"aqua",
"aquamarine",
"azure",
"beige",
"bisque",
"black",
"blanchedalmond",
"blue",
"blueviolet",
"brown",
"burlywood",
"cadetblue",
"chartreuse",
"chocolate",
"coral",
"cornflowerblue",
"cornsilk",
"crimson",
"cyan",
"darkblue",
"darkcyan",
"darkgoldenrod",
"darkgray",
"darkgreen",
"darkgrey",
"darkkhaki",
"darkmagenta",
"darkolivegreen",
"darkorange",
"darkorchid",
"darkred",
"darksalmon",
"darkseagreen",
"darkslateblue",
"darkslategray",
"darkslategrey",
"darkturquoise",
"darkviolet",
"deeppink",
"deepskyblue",
"dimgray",
"dimgrey",
"dodgerblue",
"firebrick",
"floralwhite",
"forestgreen",
"fuchsia",
"gainsboro",
"ghostwhite",
"gold",
"goldenrod",
"gray",
"green",
"greenyellow",
"grey",
"honeydew",
"hotpink",
"indianred",
"indigo",
"ivory",
"khaki",
"lavender",
"lavenderblush",
"lawngreen",
"lemonchiffon",
"lightblue",
"lightcoral",
"lightcyan",
"lightgoldenrodyellow",
"lightgray",
"lightgreen",
"lightgrey",
"lightpink",
"lightsalmon",
"lightseagreen",
"lightskyblue",
"lightslategray",
"lightslategrey",
"lightsteelblue",
"lightyellow",
"lime",
"limegreen",
"linen",
"magenta",
"maroon",
"mediumaquamarine",
"mediumblue",
"mediumorchid",
"mediumpurple",
"mediumseagreen",
"mediumslateblue",
"mediumspringgreen",
"mediumturquoise",
"mediumvioletred",
"midnightblue",
"mintcream",
"mistyrose",
"moccasin",
"navajowhite",
"navy",
"oldlace",
"olive",
"olivedrab",
"orange",
"orangered",
"orchid",
"palegoldenrod",
"palegreen",
"paleturquoise",
"palevioletred",
"papayawhip",
"peachpuff",
"peru",
"pink",
"plum",
"powderblue",
"purple",
"red",
"rosybrown",
"royalblue",
"saddlebrown",
"salmon",
"sandybrown",
"seagreen",
"seashell",
"sienna",
"silver",
"skyblue",
"slateblue",
"slategray",
"slategrey",
"snow",
"springgreen",
"steelblue",
"tan",
"teal",
"thistle",
"tomato",
"turquoise",
"violet",
"wheat",
"white",
"whitesmoke",
"yellow",
"yellowgreen",
"transparent",
"TRANSPARENT",
"",
"inherit",
"INHERIT",
"KQ@m?-ldxNK{zH+(FL_owz>YNH^]",
"aj9c)r+J&3)0E,- Lzv6K6nT@6?I}BY^\\g",
"Cf}qJN|3D>m:^[7B*fhu>lC[B3\"T_-F",
"h<s1pMd-8H[,\\caWH1oW3%M",
"z{y]VZj)%)ZV<OTbO9\\Nbc|YL}4BI<DlUBk$EV`EbN}x/@SDN0(",
"PSX2Ol8@",
"d+gYXKUM'&D'S]o<9T\\:hj_i!|l!)e6R4Bo)-(965rlK\"K01C68pgkJ] fx?kjT.O&sY4",
"[4\"tk)a/v17?0W!F`AyI[='2~;:DF6I>_<O$ he213u",
"F|r9T&N69MWq3Jrj6",
"dYR7}n&:Rq[J",
"M;Z]r@(R([6aT`<sN?uO'2Kb~3U\\\\tQUDxLN1f/D(,Q0w|K;,t`,tQ~[W/c!uQg)d|:D\\U33!DK&d*C`Zc'U#",
"kV)DKUb~h{SQCM;T*g2!Rj?>Sl=jY;3W9M{Fliu!=>tDY]5",
"y>X\\kKN|~=J+7Pqp|%9R!nZ,@>mUW9<o;|02LV<fxihsBSKVaTdcae",
"Q>jc|/:#qwzHL`lL%e~DbhQ+d^tpf9sx%o)jC1Nm}`G;rT2jo+M$=$?BC'|O^]hW^BBo_J->bWG1",
"OIxA\\5HB7g3Rv;PD)z?jGe?<x`4~9&D9dSDP=ilUauI'qb",
"aND[Al/^#;n'|V\"Vl$bh5\\G#{%y4#\\W0:ZgXe73ZuXrWcL4gr|B7,ijZZi{p)M+R9{C/&249G",
"7xK-d6Tx]BU|T,DY.qCwusmV%Ksset",
"I=UwM''S",
"w|_;Qw(R:>Clf[#3JFr_+?'1D&}WaY_xaRyTpwio>C;Pjf/kIW{]KK:R&ARiP=_g_UqRVvFKG(OQo6y'wF]Fc",
"G:",
"+XZ%s7L3FmGFn]Y!J;.vpAUoGU,&WY8eQeGWW?Jq7ANeM}[?gsV) H\\@{8z_t$oS(_jSq]|9?W*sG%' (d%",
"*P\"?'?NHA \\!{.S=+LD8Ltr^'=,$4uQ=tVL/T_b6m!PJ8*s*v`;6kp(+8i.}T!9p6{",
"_@(w<\\DjMk c8/\"/ifJNT_2R>V'}{&72C2+7otvd,$M@Yqc)L=O.muEp28m&AY",
"J!M#$z|n:+:6@7n*v)UCbkVp0;\"1#}o:i4B9oh=%'I",
"0",
"Krq?xAul2cRe&`*Fg2)bV/r>oJ`Z(ae,z%+`E@VkWH&`-jMZ<UW~jxDek;^j2\\Uq;C,Ss",
"#b\\l~=y5H=#Jy(6FwH5]jU;6D",
"YO|tw;`E_'G<d~juVPCla%K]q x\"oA-aW|Y@P$_$",
"}rI\\5x724b29MEauSSX&!AT[J1ce?,rtLAA8X",
"hlo8jd$D-dI=I#Be:BATkZPR~%Vfe0g_Xw^+wwhHQhC1;sn+P<b&J:~HfxVBX}9b/#HHPS",
"+#[?UFGUVFn0Zn7yE#TEo{FV\\{6*+s+a=fR",
"lhv.f!ENs~)?5)z:1^i@BQ|ol}9Cnkw&yV.PPx |y]@,?IL]0L_# b1'wl-]",
"&DhZ!g%v.sF}4NoP~4<vKpaM0[12!2K!ziYC3`505I*D*J6k\\skbXJ}44J#4y2",
"oK][N&iIV\"AEs3rIT-::L3&J^Sn42_J2yL= 2xI4o!b\"#2JiAt=",
"^c;C^{0wD%|y~Z1X'z\\o^gI8L=@2^p3g/L6G?]Nuif;Zf15dF`IPt8",
"62t-!*`U\\l%BFxi5B~[^~G!}h]DtXrd}y}af3",
"?N5d9ydHPi?IhwU=41'",
"GSZeLtA3YahI@oLy/6vT_[B`[PRZ1^.(n8`,8TyqVoCzMd!=9 e",
"Yck5`_#NgS",
"K9?z&o",
"Isl2>%RB8T+,9?B{~A2{fEb[%",
"&fV(`<ha/(T7J&X\\{YHt+5 =>%SaJ&W0_j]]\"",
">!sQ/IYU\"Ikc\\ei;HlCcVJ\":G2/m]h1,GvOmxFOOvTUHjHu:LWE\"QU=) ",
"7Fyx#>\"(\"N",
"MO6\"Hd2H]r8BJ}z)%J18b<VZ5lrhI",
"BGQ|tqdwj4};#x@?%ka[`DwgFWg*J+q/}]-\\\\-y#T",
"zZ=JrTPxh}.(%frt58Cy=C4(*,4]:Gnz5(~iv4@u4re~6yp:zbU0(o.S+qd9eB]A5",
"n]V3}^{9O<0cO\"rtglDO4Wc)_7Nu_JnK2EBbzRMV3b.Mj\"$9#,+-T\"N=7iPdD F<9_YWw3ZN*V;??*8VN8z?^MXi",
"fGRSl*i>^*uy|c;5B}tKXu>5hZX:>CB{oWIrxE8@B/f{:u9]:bLO0/ZWeHoNfCc|kSh{/fXs9Y:UKaJ95vFFtB2Y",
".&-4UOcxR\"Tbgc--@& hoUavCcQW^^fT}:I(d%o}J2t*BRA1{YGXB9>AYu^Bv#rEu`pN65_-r.IQD.?Cc/B({YtK[2KMmVOC3*2J",
"H3<MOq'81C#\\nUjQc  xlsF@c2R<e);T~G]^N0_*M<j!jub~k,mgZ(.>GERhwS;kmmKC?1l} qQ&zcXK?g)S OmF^=E^TlTC)/",
"8\\5tFz:sod",
"ILUc5c!'K.7.P&=S,pSYB{",
"%\\(6.jC\"C4\\2{TYdx,Ln^",
"tL3HnWq q!'k!#R@2bkvIAI]hk)64-i?uk]oK(nQiKg$`m6tx|#l>m1bEK^&.X[o'b\"\\',A16n>9ZAPI9>",
"{#Mn0,S~a\"!gw",
"dv+6'7Dd)fz",
"9o1*Q:{6#3f uHx\"_",
".43_zr}X+BtruMV!H!xw 8!9I_}zlBT3W52)rh,9ngeu0o[V_s*z'8x9*yjF!",
"y3Nm`>P:seF'V'?+<={oU5QQ",
".#Os_jv,\"@-zsY8j'POOYnY?0ONn?i#d4tqp?~.OF#VC.=<t<+feuf{#@O7lXC@+#t_uKGue%Dk9z",
"0Ep=zwydU-V<)9<9\\`[4,d^B&Gbo#'HTSEC;qU&1| ocNd69#{nbmYJMlj6Mfs3`w&pc(poie *ZOJIp7%}cVnfml",
"746}e(rye4lT7#B.Or8j1->Xs@o8f0}/e>uvvkNS[3UC2F]#[>^f74jxoo&9{^ED#,CV\"k'0|xI",
".D1{.:9gHW}]36RlUQ?!-\"0dn:+(/e@b)|'B",
"!Nh9xY2y?5ku5f}%]JWw~OfeYdcgI#It` T)3VzYl5gChve[I'rEqyJ \"@(z.Y%fEJ:9k(",
"J^$^L8:qI=yrerd0kxJBEwby6[>9[) NpqN%)h",
"l&\\K{s9]aL(,dX?B\"1g1*@9BY~=UO+AWvVRI;ar4p8Tsy~Qk-=x(yp38:j|g'5H~e4",
"(]Pe,1fTS_P%[xY96#,rwQz=P?z",
")PC#?b'&|maJBt.6Izz%vV5e!9Sy|G!*Q",
"pzAG^@/J&)Dc}GP*xO]ezW:*PV|Dt^fF<8GgF_1i=A\"@>nt?yOa|zS<8`/;uY~^ozjvX!K#%us!>2IYITh]Zy<^dq?&\"nqV]ZdZ",
"T<xz*/H&}36<(4E^/Z1m<#_G0R:=qX?1`*6Y&r'SIO]9OR;m5-Zq?PU^jvLKPLW2wPf",
")CLhmCI%TwB+t:h.@Vp-#({d0W/R_(^f34LC=V)A",
"yCT_hl%fUL<T&e;ePsT.pJG@|jO,[pN]]Q<Yu=Lp6X6&$Ka#",
" $4X",
"rp",
"t^(*U<Qdi$!vyg5D\"yQZd<K6<Qptsvzi~D}.Uf?P$E>}t8VP^;3Tamv2Z$1<",
",",
"EGCXZ_{eqIwfG-o7o\\\"!ZWTPLd,U-k$Cz]%:vNWdo}vDh!ONtM>mMP{/Cg+2<.J\"a*n#Rtnha",
"LQ3sIr7Q_ wSD7Zzv$-vxr|3P`pas0#Ze/---{PSwJ3{!a*[k'nFgC\"W+@4URi?qJk&Tt@`abNms40#A^XcAt}",
"^i356'hX\"hKaZGfTZ|C@#}b3LGz1\">qcH{L8{Fs?O5%:EqQWuro_lc=]gWLVR\\~!J\"[>,H",
"i{*q<O{Dt^n|FY3,FG>WIRqPH.os$9^P=|yA9?P;MOw;VBwZ^>K=\"%J9SBlv%0+o5k73rW!`l_",
"-PCM&!G~o#Za^s&)qJELr\\P^\\={_xTFp:%@JF.PeRX\\7b8K",
"V_sFicB+wx5",
"\"X^\\d}b9.W.2\"O!yAL21\"Ny5:)=Q3*D|TAzzr^0Yflzjdc!p*.yW,B",
"kZ`wCP>9Bq5S!r!Vi|Uy/C&H[kz/f^{(Z[OGw'S0\"",
"Jn%1^rUnNB|%=q%^v*bN|I40}#Htn{G!#~CNAN0KvZcB>Ita(,n",
"-PlhE[^J55Ui",
"z`h`uQJ{J",
"eV\\q5Q4o@Y*,IRMcnpqj5>Id\\yBe?pKH3uF&c<c}:E9[uaH$ 8dXCmI+!C'q@PkE<NVRq~GRW<tfyt/i@%dwI&rL",
"UGzLF+o3)Ezs=nMxqd^\"=q.Ik}Tk2I`X)R8]Zmy/WQp,|]TdbP)5  J+#Hm6SmWtQ+h?.MQ1W#oyp\\F,'JL>rLtjiHOA",
"&joOSw7XZVvSt4ZCT*:aq:3ns!v|r);~7gN8'_D",
"Y<q|Mhn5Nrcb+dR=10pQF5]r@/*7P`79w/htSm2,C~1&sUW{N@v:t9d;HPG&xrI\"YD;V9Y$'g,W'J=GV3,YK",
"Gx&#{;]l/?[{SyX`kTeo",
"30PU7@<'58.hRWsJTa9L.hVQ8}7=$}ih4|$Y*9z3[aooT!]}+>b{1JH^.jjEU{,dAXSCbtEh6",
"%2~x8=A!RW@8N/`hQz`)gl}1DOU9{>Ie'L> 4e]m;kt =isEQ(\\TeI7hWgK-K! p^K'\":3;dxTLO",
"\\ ):{Woay[4",
"\\{Ih&}*8^x6@`V@DZB`rSuhYm4i@TW^t9Hx[^`IVumjXc1vA\"~wt8^Jf:US6Z\\xaS&",
"lo $6<EP|=gAEpd\\M6YDg\"*0m",
/*
"ActiveBorder",
"ActiveCaption",
"AppWorkspace",
"Background",
"ButtonFace",
"ButtonHighlight",
"ButtonShadow",
"ButtonText",
"CaptionText",
"GrayText",
"Highlight",
"HighlightText",
"InactiveBorder",
"InactiveCaption",
"InactiveCaptionText",
"InfoBackground",
"InfoText",
"Menu",
"MenuText",
"Scrollbar",
"ThreeDDarkShadow",
"ThreeDFace",
"ThreeDHighlight",
"ThreeDLightShadow",
"ThreeDShadow",
"Window",
"WindowFrame",
"WindowText",
*/
"currentColor",
"0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f",
"00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f",
"#000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f",
"#0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000f",
"#",
"#1",
"#12",
"#123",
"#1234",
"#12x",
"abc",
"123",
"#0ab0cd0ef",
"#0ab0cd0e",
"#0ab0cd0",
"#0ab0cdaef",
"#0ab0cdae",
"#0ab0cda"
];

var references = [
"#113355",
"#001350",
"#135000",
"#001350",
"#013500",
"#000135",
"#123456",
"#002356",
"#124500",
"#002356",
"#013460",
"#001245",
"#f0f8ff",
"#f0f8ff",
"#a0ce00",
"#a0ce00",
"#0ace0e",
"#a0ebe0",
"#00cee0",
"#faebd7",
"#00ffff",
"#7fffd4",
"#f0ffff",
"#f5f5dc",
"#ffe4c4",
"#000000",
"#ffebcd",
"#0000ff",
"#8a2be2",
"#a52a2a",
"#deb887",
"#5f9ea0",
"#7fff00",
"#d2691e",
"#ff7f50",
"#6495ed",
"#fff8dc",
"#dc143c",
"#00ffff",
"#00008b",
"#008b8b",
"#b8860b",
"#a9a9a9",
"#006400",
"#a9a9a9",
"#bdb76b",
"#8b008b",
"#556b2f",
"#ff8c00",
"#9932cc",
"#8b0000",
"#e9967a",
"#8fbc8f",
"#483d8b",
"#2f4f4f",
"#2f4f4f",
"#00ced1",
"#9400d3",
"#ff1493",
"#00bfff",
"#696969",
"#696969",
"#1e90ff",
"#b22222",
"#fffaf0",
"#228b22",
"#ff00ff",
"#dcdcdc",
"#f8f8ff",
"#ffd700",
"#daa520",
"#808080",
"#008000",
"#adff2f",
"#808080",
"#f0fff0",
"#ff69b4",
"#cd5c5c",
"#4b0082",
"#fffff0",
"#f0e68c",
"#e6e6fa",
"#fff0f5",
"#7cfc00",
"#fffacd",
"#add8e6",
"#f08080",
"#e0ffff",
"#fafad2",
"#d3d3d3",
"#90ee90",
"#d3d3d3",
"#ffb6c1",
"#ffa07a",
"#20b2aa",
"#87cefa",
"#778899",
"#778899",
"#b0c4de",
"#ffffe0",
"#00ff00",
"#32cd32",
"#faf0e6",
"#ff00ff",
"#800000",
"#66cdaa",
"#0000cd",
"#ba55d3",
"#9370db",
"#3cb371",
"#7b68ee",
"#00fa9a",
"#48d1cc",
"#c71585",
"#191970",
"#f5fffa",
"#ffe4e1",
"#ffe4b5",
"#ffdead",
"#000080",
"#fdf5e6",
"#808000",
"#6b8e23",
"#ffa500",
"#ff4500",
"#da70d6",
"#eee8aa",
"#98fb98",
"#afeeee",
"#db7093",
"#ffefd5",
"#ffdab9",
"#cd853f",
"#ffc0cb",
"#dda0dd",
"#b0e0e6",
"#800080",
"#ff0000",
"#bc8f8f",
"#4169e1",
"#8b4513",
"#fa8072",
"#f4a460",
"#2e8b57",
"#fff5ee",
"#a0522d",
"#c0c0c0",
"#87ceeb",
"#6a5acd",
"#708090",
"#708090",
"#fffafa",
"#00ff7f",
"#4682b4",
"#d2b48c",
"#008080",
"#d8bfd8",
"#ff6347",
"#40e0d0",
"#ee82ee",
"#f5deb3",
"#ffffff",
"#f5f5f5",
"#ffff00",
"#9acd32",
"transparent",
"transparent",
"transparent",
"#00e000",
"#00e000",
"#0df000",
"#0000b0",
"#007b30",
"#008001",
"#004b00",
"#002080",
"#099600",
"#0120e2",
"#f00630",
"#d00000",
"#200000",
"#0c00d0",
"#07900a",
"#db020b",
"#7000a0",
"#b00c02",
"#d6d000",
"#000000",
"#1d00f0",
"#000000",
"#a00000",
"#00b600",
"#007028",
"#00b0b9",
"#000000",
"#f00000",
"#b00005",
"#0ea000",
"#72000a",
"#ba0c00",
"#0fe000",
"#50000b",
"#000c44",
"#ae4202",
"#00005d",
"#0000af",
"#5d0041",
"#a00000",
"#c00000",
"#090000",
"#0209fe",
"#a00500",
"#c0100e",
"#7f0000",
"#600000",
"#d00d00",
"#0004d9",
"#c00000",
"#0500ff",
"#0000c3",
"#200000",
"#80f00d",
"#0c7000",
"#00c4d0",
"#a000a0",
"#00a000",
"#d0070f",
"#900600",
"#002090",
"#30ef00",
"#0000e0",
"#b000c0",
"#80c200",
"#0900e0",
"#0005e0",
"#800b00",
"#b0a000",
"#e00000",
"#000690",
"#d00000",
"#010020",
"#0000c0",
"#000060",
"#004000",
"#000000",
"#000da0",
"#000000",
"#00d000",
"#000fa0",
"#0c5000",
"#000030",
"#a0007b",
"#00cb05",
"#023000",
"#9bc000",
"#b000b0",
"#00e055",
"#000000",
"#0dc0d0",
"#20600a",
"#70a070",
"#f50000",
"#0000e0",
"#a900cb",
"#0000d0",
"#000a40",
"#d00060",
"#000ad0",
/*
"ActiveBorder",
"ActiveCaption",
"AppWorkspace",
"Background",
"ButtonFace",
"ButtonHighlight",
"ButtonShadow",
"ButtonText",
"CaptionText",
"GrayText",
"Highlight",
"HighlightText",
"InactiveBorder",
"InactiveCaption",
"InactiveCaptionText",
"InfoBackground",
"InfoText",
"Menu",
"MenuText",
"Scrollbar",
"ThreeDDarkShadow",
"ThreeDFace",
"ThreeDHighlight",
"ThreeDLightShadow",
"ThreeDShadow",
"Window",
"WindowFrame",
"WindowText",
*/
"#c0e000",
"#0000f0",
"#000000",
"#0000f0",
"#000000",
"#000000",
"#010000",
"#010200",
"#112233",
"#123400",
"#010200",
"#0a0b0c",
"#010203",
"#abcdef",
"#abcde0",
"#abcd00",
"#0a0cae",
"#0a0cae",
"#0a0ca0"
];

var todos = {
" #135": true,
"#135 ": true,
" #135 ": true,
"#   135": true,
" #123456": true,
"#123456 ": true,
" #123456 ": true,
"#   123456": true,
" aliceblue": true,
"aliceblue ": true,
"  aliceblue   ": true,
"H3<MOq'81C#\\nUjQc  xlsF@c2R<e);T~G]^N0_*M<j!jub~k,mgZ(.>GERhwS;kmmKC?1l} qQ&zcXK?g)S OmF^=E^TlTC)/": true,
" $4X": true,
"UGzLF+o3)Ezs=nMxqd^\"=q.Ik}Tk2I`X)R8]Zmy/WQp,|]TdbP)5  J+#Hm6SmWtQ+h?.MQ1W#oyp\\F,'JL>rLtjiHOA": true
};

var table0 = document.getElementById("table0");
var table1 = document.getElementById("table1");
var cs0 = document.defaultView.getComputedStyle(table0);
var cs1 = document.defaultView.getComputedStyle(table1);
var result;
var reference;
var log = "";
var len = tests.length;
is(tests.length, references.length, "array length mismatch");
for (var i = 0; i < len; ++i) {
  table0.setAttribute("bgColor", tests[i]);
  table1.style.backgroundColor = references[i];
  ((tests[i] in todos) ? todo_is : is)(
     cs0.getPropertyValue("background-color"),
     cs1.getPropertyValue("background-color"),
     "html color '" + tests[i] + "' should match '" + references[i] + "'");
}
</script>

</script>
</pre>
</body>
</html>
